Incluindo Autenticação e Iniciando permissão
Nova versão de banco de dados

+++ app/Http/Controllers/BlocoController.php
@@ -11,57 +11,64 @@ class BlocoController extends Controller
 {
     public function index()
     {
-        $blocos = Bloco::all();
-        return view('blocos.bloco-listar', compact('blocos'));
+        Log::info('userlogado', [auth()->user()]);
+        //verifica se tem usuário logado e se é admin ou coordenador
+        if(Auth::check() && (auth()->user()->hasRole('admin') || auth()->user()->hasRole('coordenador'))){
+            $blocos = Bloco::all();
+            return view('blocos.bloco-listar', compact('blocos'));
+        }else{
+            return redirect()->route('login');
+        }
     }
 
     public function create()
     {
-        // if(Auth::check()){
+        if(Auth::check() && (auth()->user()->hasRole('admin') || auth()->user()->hasRole('coordenador'))){
             return view('blocos.bloco-cadastrar');
-        // }else{
-            //return redirect()->route('blocos.index');
-        //}
+        }else{
+            return redirect()->route('login');
+        }
     }
 
     public function store(Request $request)
     {
         Log::info('request', [$request]);
-        //if(Auth::check()){
+        if(Auth::check() && (auth()->user()->hasRole('admin') || auth()->user()->hasRole('coordenador'))){
             Bloco::create($request->all());
-        //}
-        return redirect()->route('blocos.index');
-    }
-
-    public function show(Bloco $bloco)
-    {
-        return view('blocos.show', compact('bloco'));
+            return redirect()->route('blocos.index');
+        }else{
+            return redirect()->route('login');
+        }
     }
 
     public function edit(Bloco $bloco)
     {
-        //if(Auth::check()){
-           return view('blocos.bloco-editar', compact('bloco'));
-        //}else{
-           // return redirect()->route('blocos.index');
-        //}        
+
+        if(Auth::check() && (auth()->user()->hasRole('admin') || auth()->user()->hasRole('coordenador'))){
+            return view('blocos.bloco-editar', compact('bloco'));
+        }else{
+            return redirect()->route('login');
+        }
+
     }
 
     public function update(Request $request, Bloco $bloco)
     {
-        //if(Auth::check()){
+        if(Auth::check() && (auth()->user()->hasRole('admin') || auth()->user()->hasRole('coordenador'))){
            $bloco->update($request->all());
-        //}else{
             return redirect()->route('blocos.index');
-        //}    
+        }else{
+            return redirect()->route('login');
+        }   
     }
 
     public function destroy(Bloco $bloco)
     {
-        //if(Auth::check()){
+        if(Auth::check() && (auth()->user()->hasRole('admin') || auth()->user()->hasRole('coordenador'))){
             $bloco->delete();
-        // }else{
             return redirect()->route('blocos.index');
-        // }    
+        }else{
+            return redirect()->route('login');
+        }    
     }
 }
